requirements.txt
# Paper Search Assistant - Python Dependencies

# Core search libraries
findpapers>=0.3.0
arxiv>=2.1.0
pymed>=0.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional (for advanced features)
# paper-qa>=0.0.20  # Uncomment if using PaperQA
